Basic DSL #let operation to set variables and reload before eval
let(:var) {:result} will create a variable named "var" that will
always evaluate to the result of the passed block.  Any source
files detected as defining a piece of the result will be hard
reloaded prior to evaluation.  This allows for very dynamic
tinkering in a pry or irb session.